+++ app/Http/Controllers/NoteController.php
@@ -62,12 +62,12 @@ class NoteController extends Controller
         $note->tags;
         // does the note exist? If not, return error
         if(!$note)
-        return redirect('notes/index')->withErrors(['message', 'That note does not exist!']);
+            return redirect('notes/index')->withErrors(['That note does not exist!']);
 
         // Get all note and their tags if the tag ID is in our tags-array
         $tags = [];
         foreach($note->tags as $tag)
-        $tags[] = $tag->id;
+            $tags[] = $tag->id;
 
         // Get all notes that have at least one of this note's tags
         $relatedNotes = Note::whereHas('tags', function($query) use($tags) {
@@ -105,6 +105,36 @@ class NoteController extends Controller
         //array_multisort($relatedNotes, "SORT_DESC", );
         $relatedNotes = Collection::make($relatedNotes)->sortBy(function($value) { return $value->count; }, SORT_REGULAR, true)->all();
 
+        // Now it can be, that some people are just ignoring the tag functionality
+        // which is fine. But we then need another algorithm to search notes:
+        // heuristic text analysis!
+        if(count($note->tags) == 0 || count($relatedNotes) < 5) {
+            $relatedNotes = $this->textAnalysis($note);
+
+            // Now retrieve all related Notes
+            $tmp = new Collection();
+            $maxCount = 0;
+            foreach($relatedNotes as $id)
+            {
+                $n = Note::find($id['id']);
+                $n->count = $id['relevancy'];
+                $tmp->push($n);
+
+                if($id['relevancy'] > $maxCount) {
+                    $maxCount = $id['relevancy'];
+                }
+            }
+            $relatedNotes = Collection::make($tmp)->sortBy(function($value) { return $value->count; }, SORT_REGULAR, true)->all();
+            // Remove this note to reduce inception level
+            $thisNoteIndex = -1;
+            foreach($relatedNotes as $index => $n)
+            {
+                if($note->id == $n->id)
+                    $thisNoteIndex = $index;
+            }
+            unset($relatedNotes[$thisNoteIndex]);
+            $relatedNotes = array_values($relatedNotes);
+        }
         // Now retrieve IDs and title of all linked notes
         $linkedNotes = $note->notes;
         $mainID = $note->id;
@@ -326,5 +356,64 @@ class NoteController extends Controller
         $note->delete();
 
         return redirect(url('/notes/index'));
+    }
+
+    function textAnalysis(Note $note)
+    {
+        // For a text analysis we need to reduce this note and
+        // reduce all possible related notes to a small index.
+        // TODO: Implement an indexing function.
+
+        // explode note contents and title by words
+        $title = explode(" ", $note->title);
+        $tmp = strip_tags(preg_replace('/[^a-zA-Z0-9[:space:]\p{L}]/u', '', $note->content));
+        $tmp = str_replace(["\n", "\r"], " ", $tmp);
+        $content = explode(" ", $tmp);
+        // Remove duplicates and reset array keys to Zero-based
+        $content = array_values(array_unique($content));
+        $suggestedNotes = [];
+        $rel = [];
+        // Now suggest tags for all words with > 3 letters
+        for($i = 0; $i < count($content); $i++)
+        {
+            if(strlen($content[$i]) > 3)
+            {
+                // I guess above 300+ notes this will be one of the slowest
+                // functions in this whole app. Let's see if it breaks the
+                // 60sec wall of the server ...
+                $notes = Note::where('content', 'LIKE', '%'.$content[$i].'%')->get();
+                if(count($notes) > 0)
+                {
+                    foreach($notes as $index => $n)
+                    {
+                        // Compare the arrays to determine the relevancy
+                        $suggestedNotes[$index] = $n->id;
+
+                        $tmp = strip_tags(preg_replace('/[^a-zA-Z0-9[:space:]\p{L}]/u', '', $n->content));
+                        $tmp = str_replace(["\n", "\r"], " ", $tmp);
+                        $suggestContent = explode(" ", $tmp);
+                        // Remove duplicates and reset array keys to Zero-based
+                        $suggestContent = array_values(array_unique($suggestContent));
+
+                        $relevancy = count(array_intersect($content, $suggestContent));
+                        $rel[$index] = $relevancy;
+                    }
+                }
+            }
+        }
+
+        // Remove potential duplicate notes
+        $suggestedNotes = array_unique($suggestedNotes);
+        // now write the index of relevancy into the remaining $suggestedNotes
+        unset($tmp);
+        $tmp = [];
+        foreach($suggestedNotes as $index => $id)
+        {
+            $tmp[$index]['id'] = $id;
+            $tmp[$index]['relevancy'] = $rel[$index];
+        }
+        $suggestedNotes = $tmp;
+
+        return $suggestedNotes;
     }
 }